requirements.txt
langchain==0.3
numpy
pandas
python-dotenv==1.0.1
langchain_huggingface